Updating to latest version

+++ task-force-frontend/src/app/home/home.component.ts
@@ -8,25 +8,29 @@ import { RouterLink } from '@angular/router';
   styleUrls: ['./home.component.css']
 })
 export class HomeComponent implements OnInit {
+
   addNew = false;
-  constructor(private taskService: TaskService) { 
+
+  constructor(private taskService: TaskService) {
   }
 
   ngOnInit(): void {
-    
-  }
-  
-  addNewTask () {
-    return this.addNew = true;
+
   }
 
-  deleteTasks () {
-    this.taskService.deleteAllTasks().subscribe(response => {
-      console.log(`Response = ${response}`);
-      alert (`All tasks have been deleted!`)
-    },
-    err => {
-      console.log(err);
-    });
+  toggleCreate() {
+    return this.addNew = !this.addNew;
+  }
+
+  deleteTasks() {
+    if (confirm("Do you Really want to DELETE ALL tasks!")) {
+      this.taskService.deleteAllTasks().subscribe(response => {
+        console.log(`Response = ${response}`);
+        alert(`All tasks have been deleted!`)
+      },
+        err => {
+          console.log(err);
+        });
+    }
   }
 }
